Changed VS Compile Versions, minor conversion error corrections in enemy.cpp

+++ Platformer/Src/Enemies/Enemy.cpp
@@ -5,7 +5,7 @@ Enemy::Enemy(const sf::Vector2i & spawn_position)
 	m_defaultHorizontalSpeed(0.3f),
 	m_isGoingLeft(true)
 {
-	m_mainObject.setPosition(sf::Vector2f(spawn_position.x * 16, spawn_position.y * 16));
+	m_mainObject.setPosition(sf::Vector2f(float(spawn_position.x * 16), float(spawn_position.y * 16)));
 	m_mainObject.setFillColor(sf::Color::Red);
 	m_mainObject.setSize(sf::Vector2f(16, 16));
 }
@@ -17,8 +17,8 @@ void Enemy::Update(const std::vector<std::string>& map)
 	else
 		m_mainObject.move(sf::Vector2f(m_defaultHorizontalSpeed, 0));
 
-	int x = m_mainObject.getPosition().x / 16;
-	int y = m_mainObject.getPosition().y / 16;
+	int x = int(m_mainObject.getPosition().x) / 16;
+	int y = int(m_mainObject.getPosition().y) / 16;
 
 	if (m_isGoingLeft)
 	{
